+++ pkg/dartdev/lib/src/commands/run.dart
@@ -422,10 +422,14 @@ class RunCommand extends DartdevCommand {
             return errorExitCode;
           }
         } else if (await builder.hasHooks()) {
-          final assetsYamlFileUri = await progress(
-            'Running build hooks',
-            builder.compileNativeAssetsJitYamlFile,
-          );
+          final verbosity = args.option('verbosity')!;
+          final showProgress = verbosity != Verbosity.error.name;
+          final assetsYamlFileUri = await (showProgress
+              ? progress(
+                  'Running build hooks',
+                  builder.compileNativeAssetsJitYamlFile,
+                )
+              : builder.compileNativeAssetsJitYamlFile());
           if (assetsYamlFileUri == null) {
             log.stderr('Error: Running build hooks failed.');
             return errorExitCode;